Add widget settings helper and rename widgetAccessHelpers
