+++ src/components/Site/SiteFrameEmbed.js
@@ -1,4 +1,4 @@
-import React, { Fragment } from 'react'
+import React from 'react'
 import { Routes, Route } from 'react-router-dom'
 import PropTypes from 'prop-types'
