+++ src/components/WidgetAreaEdit/WidgetAreaEditPresentationSettingsPane.js
@@ -16,7 +16,7 @@ import {
   getWidgetLabel,
   isUUID,
 } from 'common/utils/stringHelpers'
-import { isEditableWidgetType } from 'common/utils/widgetAccessHelpers'
+import { isEditableWidgetType } from 'common/utils/widgetHelpers'
 import getWidgetIcon from 'common/utils/getWidgetIcon'
 import { WIDGET_WOTD } from 'common/constants'
